src/requirements.txt
numpy
python-dotenv
pre-commit
